Make orchestra.console a subpackage
