+++ teuthology/orchestra/console/__init__.py
@@ -1,0 +1,5 @@
+import physical
+import virtual
+
+PhysicalConsole = physical.PhysicalConsole
+VirtualConsole = virtual.VirtualConsole
