+++ teuthology/orchestra/test/test_console.py
@@ -88,7 +88,7 @@ class TestPhysicalConsole(TestConsole):
 
     def test_spawn_log_conserver(self):
         with patch(
-            'teuthology.orchestra.console.psutil.subprocess.Popen',
+            'teuthology.orchestra.console.physical.psutil.subprocess.Popen',
             autospec=True,
         ) as m_popen:
             m_popen.return_value.pid = 42
@@ -107,7 +107,7 @@ class TestPhysicalConsole(TestConsole):
 
     def test_spawn_log_ipmi(self):
         with patch(
-            'teuthology.orchestra.console.psutil.subprocess.Popen',
+            'teuthology.orchestra.console.physical.psutil.subprocess.Popen',
             autospec=True,
         ) as m_popen:
             m_popen.return_value.pid = 42
@@ -126,7 +126,7 @@ class TestPhysicalConsole(TestConsole):
 
     def test_spawn_log_fallback(self):
         with patch(
-            'teuthology.orchestra.console.psutil.subprocess.Popen',
+            'teuthology.orchestra.console.physical.psutil.subprocess.Popen',
             autospec=True,
         ) as m_popen:
             m_popen.return_value.pid = 42
@@ -146,7 +146,7 @@ class TestPhysicalConsole(TestConsole):
 
     def test_get_console_conserver(self):
         with patch(
-            'teuthology.orchestra.console.psutil.subprocess.Popen',
+            'teuthology.orchestra.console.physical.psutil.subprocess.Popen',
             autospec=True,
         ) as m_popen:
             m_popen.return_value.pid = 42
@@ -155,7 +155,7 @@ class TestPhysicalConsole(TestConsole):
             cons = self.klass(self.hostname)
         assert cons.has_conserver is True
         with patch(
-            'teuthology.orchestra.console.pexpect.spawn',
+            'teuthology.orchestra.console.physical.pexpect.spawn',
             autospec=True,
         ) as m_spawn:
             cons._get_console()
@@ -165,7 +165,7 @@ class TestPhysicalConsole(TestConsole):
 
     def test_get_console_ipmitool(self):
         with patch(
-            'teuthology.orchestra.console.psutil.subprocess.Popen',
+            'teuthology.orchestra.console.physical.psutil.subprocess.Popen',
             autospec=True,
         ) as m_popen:
             m_popen.return_value.pid = 42
@@ -174,7 +174,7 @@ class TestPhysicalConsole(TestConsole):
             cons = self.klass(self.hostname)
         assert cons.has_conserver is True
         with patch(
-            'teuthology.orchestra.console.pexpect.spawn',
+            'teuthology.orchestra.console.physical.pexpect.spawn',
             autospec=True,
         ) as m_spawn:
             cons.has_conserver = False
@@ -184,7 +184,7 @@ class TestPhysicalConsole(TestConsole):
 
     def test_get_console_fallback(self):
         with patch(
-            'teuthology.orchestra.console.psutil.subprocess.Popen',
+            'teuthology.orchestra.console.physical.psutil.subprocess.Popen',
             autospec=True,
         ) as m_popen:
             m_popen.return_value.pid = 42
@@ -193,7 +193,7 @@ class TestPhysicalConsole(TestConsole):
             cons = self.klass(self.hostname)
         assert cons.has_conserver is True
         with patch(
-            'teuthology.orchestra.console.pexpect.spawn',
+            'teuthology.orchestra.console.physical.pexpect.spawn',
             autospec=True,
         ) as m_spawn:
             cons.has_conserver = True
@@ -206,7 +206,7 @@ class TestPhysicalConsole(TestConsole):
 
     def test_disable_conserver(self):
         with patch(
-            'teuthology.orchestra.console.psutil.subprocess.Popen',
+            'teuthology.orchestra.console.physical.psutil.subprocess.Popen',
             autospec=True,
         ) as m_popen:
             m_popen.return_value.pid = 42
